Experiment with Module
Introduce a Module protocol which is both View and Reducer. Not sure if this is helpful? It's more Elm-like and puts reducer with the View. It could conflate things as reduce() could try to touch Store, etc which it shouldn't. Also, reducer composition doesn't happen in the view itself.

Some quality of life things: make `SideEffects.none` a computed property instead of func, `NoActions`, `NoEffects` to go along with `EmptyReducer` for times you don't need the type.